Gundel/routes index bug

+++ lib/widgets/routes/routeCard.dart
@@ -5,7 +5,6 @@ import 'package:climbing_gym_app/services/routeColorService.dart';
 import 'package:climbing_gym_app/services/routesService.dart';
 import 'package:flutter/material.dart';
 import 'package:climbing_gym_app/constants.dart' as Constants;
-import 'package:provider/provider.dart';
 import 'package:transparent_image/transparent_image.dart';
 import '../../locator.dart';
 
@@ -20,10 +19,22 @@ class RouteCard extends StatefulWidget {
 }
 
 class _RouteCardState extends State<RouteCard> {
-  final AppRoute route;
-  final AppUser appUser;
+  AppRoute route;
+  AppUser appUser;
   _RouteCardState(this.route, this.appUser);
   final routeColorService = locator<RouteColorService>();
+
+  @override
+  void didUpdateWidget(RouteCard oldWidget) {
+    if (route != widget.route || appUser != widget.appUser) {
+      setState(() {
+        route = widget.route;
+        appUser = widget.appUser;
+      });
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -124,40 +135,38 @@ class _RouteCardState extends State<RouteCard> {
 
   void onPressDelete() {
     final routesService = locator<RoutesService>();
-
-    showDialog(
-      context: context,
-      builder: (_) {
-        return AlertDialog(
-          title: Text(
-            'Delete Route',
-          ),
-          content: SingleChildScrollView(
-            child: ListBody(
-              children: <Widget>[
-                Text(
-                  'Would you like to delete this route?',
-                ),
-              ],
+    if (this.mounted) {
+      showDialog(
+        context: context,
+        builder: (_) {
+          return AlertDialog(
+            title: Text(
+              'Delete Route',
             ),
-          ),
-          actions: <Widget>[
-            TextButton(
-                onPressed: () => Navigator.of(context).pop(),
-                child: Text("No")),
-            TextButton(
-                onPressed: () async {
-                  bool isDeleted =
-                      await routesService.deleteRoute(this.route.id);
-                  if (isDeleted) {
-                    Navigator.of(context, rootNavigator: true).pop();
-                  }
-                },
-                child: Text("Yes")),
-          ],
-        );
-      },
-    );
+            content: SingleChildScrollView(
+              child: ListBody(
+                children: <Widget>[
+                  Text(
+                    'Would you like to delete this route?',
+                  ),
+                ],
+              ),
+            ),
+            actions: <Widget>[
+              TextButton(
+                  onPressed: () => Navigator.of(context).pop(),
+                  child: Text("No")),
+              TextButton(
+                  onPressed: () async {
+                    routesService.deleteRoute(this.route.id);
+                    Navigator.of(context).pop();
+                  },
+                  child: Text("Yes")),
+            ],
+          );
+        },
+      );
+    }
   }
 
   bool _getIsPrivileged() {
